Add logic to determine validity based on API response
When the API responds with 200 it includes the results of the validation
request, there are two main elements to be focussed on, success and
classification.
From limited testing using valid and invalid @uad.ac.uk email addresses (it
doesnt matter the domain that is provided on the invite site, it will always be
replaced with uad.ac.uk) status can have the values, Success or
MailboxDoesNotExist. Whereas classification can be Deliverable or Undeliverable.