Add case for response code 202 from API requests
This code indicates that the job was queued for completion. To handle
this case multiple addittional requests need to be made.
Firstly to be able to check job status we need the ID of the job, this
is found under `["overview"]["id"]`. On subsequent requests the ID is
appended to the `/email-validations/` endpoint.
Each time the API responds with 202 on this new endpoint, it is telling
us to check again later, i'm taking later as meaning 1 second. After a
couple cycles a 200 response is returned and the job is complete, again
checking the status and classification allows to determine the validity

- to force the API to respond with 202 for testing, you can set the waitTime
  parameter to 0